requirements.txt
aiosqlite==0.20.0
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.112.1
fastapi-sessions==0.3.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
mailjet-rest==1.3.4
Mako==1.3.5
mangum==0.19.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
paystackapi==2.1.3
pdfkit==1.0.0
platformdirs==4.2.2
pluggy==1.6.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pyzmq==26.1.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
stack-data==0.6.3
starlette==0.38.2
tornado==6.4.1
traitlets==5.14.3
trio==0.30.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
websockets==13.0.1
